% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{forward_kl_loss}
\alias{forward_kl_loss}
\title{Forward KL Loss}
\usage{
forward_kl_loss(input)
}
\arguments{
\item{input}{The output of the flow model, which should be a tensor of
dimensions \verb{[batch1, ..., batchN, d]} where \verb{batch1, ..., batchN} are the
dimensions of the batch and \code{d} is the dimension of the input. It must also
have an attribute \code{log_jacobian} containing the log determinant of the
Jacobian of the transformation.}
}
\description{
Compute the forward KL loss for a flow model.
}
\examples{
library(torch)
flow_model <- nn_sequential_conditional_flow(
  nn_affine_coupling_block(2, 0),
  nn_permutation_flow(2),
  nn_affine_coupling_block(2, 0)
)
x <- torch_randn(10, 2)
y <- flow_model(x)
loss <- forward_kl_loss(y)
}
