% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_from_conditional_flow}
\alias{generate_from_conditional_flow}
\title{Generate samples from a conditional flow model}
\usage{
generate_from_conditional_flow(model, n_samples_per_batch, conditioning)
}
\arguments{
\item{model}{A conditional flow model.}

\item{n_samples_per_batch}{The number of samples to generate for each batch
of the conditioning variable, or the total number of samples if conditioning
is not provided.}

\item{conditioning}{The conditioning variable, a torch tensor of dimensions
\verb{[batch, ...]} where \code{batch} is the dimension of the batch and \code{...} are the
dimensions of the conditioning variable.}
}
\description{
This function generates samples from a conditional flow model. If
conditioning is provided, \code{n_samples_per_batch} samples are generated for
each batch of the conditioning variable. If conditioning is not provided,
\code{n_samples_per_batch} samples are generated.
}
