% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actnorm.R
\name{nn_actnorm_block}
\alias{nn_actnorm_block}
\title{ActNorm Block}
\usage{
nn_actnorm_block(input_size)
}
\arguments{
\item{input_size}{The size of the input to the flow.}
}
\description{
An ActNorm block is a conditional flow inheriting from
\code{\link[=nn_conditional_flow]{nn_conditional_flow()}} that applies a learned per-dimension affine
transformation,
}
\details{
\deqn{y = s x + b}

where \code{s} and \code{b} are trainable vectors initialized to one and zero. The
block ignores conditioning inputs and can be inserted directly into
\code{\link[=nn_sequential_conditional_flow]{nn_sequential_conditional_flow()}}.
}
\examples{
library(torch)
flow_model <- nn_sequential_conditional_flow(
  nn_actnorm_block(2),
  nn_spline_coupling_block(2),
  nn_permutation_flow(2),
  nn_actnorm_block(2),
  nn_spline_coupling_block(2)
)

}
\seealso{
\code{\link[=nn_conditional_flow]{nn_conditional_flow()}}
}
