% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional-network.R
\name{nn_conditional}
\alias{nn_conditional}
\title{Conditional Module}
\usage{
nn_conditional()
}
\description{
A conditional module is a module that takes an additional conditioning input to
the forward pass.
}
\section{Forward method}{

The forward method should take two arguments, \code{input} and \code{conditioning}, and
return the output. Example:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{forward = function(input, conditioning) \{
  output <- ...
  output
\}
}\if{html}{\out{</div>}}
}

\seealso{
\code{\link[=nn_conditional_mlp]{nn_conditional_mlp()}} for a concrete implementation.
}
