% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{nn_coupling_transform}
\alias{nn_coupling_transform}
\title{Coupling Transform}
\usage{
nn_coupling_transform()
}
\description{
A coupling transform maps an input partition and a set of parameters to an
invertible output partition. Coupling blocks use \code{params_per_dim()} to size
their parameter networks.
}
\section{Transform methods}{

Subclasses should implement \code{params_per_dim()}, \code{split_parameters()},
\code{constrain_parameters()}, \code{forward()}, and \code{reverse()}.
}

