% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupling.R
\name{nn_dual_coupling_block}
\alias{nn_dual_coupling_block}
\title{Dual Coupling Block}
\usage{
nn_dual_coupling_block(
  input_size,
  conditioning_size = 0,
  left_size = as.integer(input_size\%/\%2),
  transform = "affine",
  f_params,
  g_params,
  ...
)
}
\arguments{
\item{input_size}{The dimension of the input.}

\item{conditioning_size}{The dimension of the conditioning input.}

\item{left_size}{The dimension of the left part of the input.}

\item{transform}{The transform to apply. Currently \code{"affine"} is supported.
A transform constructor or transform module can also be supplied.}

\item{f_params}{A conditional network returning parameters for transforming
the left part from the right part.}

\item{g_params}{A conditional network returning parameters for transforming
the right part from the transformed left part.}

\item{...}{Additional arguments passed to the transform constructor.}
}
\description{
A dual coupling block applies two single coupling transformations in
sequence: first transforming the left part from the right part, then
transforming the right part from the transformed left part.
}
\examples{
library(torch)
coupling <- nn_dual_coupling_block(4, transform = "affine")
x <- torch_randn(10, 4)
y <- coupling(x)
x_recovered <- coupling$reverse(y)

}
