% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential-flow.R
\name{nn_sequential_conditional_flow}
\alias{nn_sequential_conditional_flow}
\title{Sequential Conditional Flow}
\usage{
nn_sequential_conditional_flow(...)
}
\arguments{
\item{...}{A sequence of conditional flows, or a list of conditional flows.}
}
\description{
A sequential conditional flow is a conditional flow inheriting from
\code{\link[=nn_conditional_flow]{nn_conditional_flow()}} that applies a sequence of conditional flows, passing
the conditioning input to each flow. It is the analog of
\code{\link[torch:nn_sequential]{torch::nn_sequential()}} for conditional flows.
}
\examples{
flow_model <- nn_sequential_conditional_flow(
  nn_affine_coupling_block(10, 5),
  nn_permutation_flow(10),
  nn_affine_coupling_block(10, 5)
)

}
\seealso{
\code{\link[=nn_conditional_flow]{nn_conditional_flow()}}
}
