% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-coupling.R
\name{nn_spline_coupling_block}
\alias{nn_spline_coupling_block}
\title{Spline Coupling Block}
\usage{
nn_spline_coupling_block(
  input_size,
  conditioning_size = 0,
  left_size = as.integer(input_size%/%2),
  num_bins = 8,
  f_params,
  g_params,
  tail_bound = 3,
  min_bin_width = 0.001,
  min_bin_height = 0.001,
  min_derivative = 0.001
)
}
\arguments{
\item{input_size}{The dimension of the input.}
\item{conditioning_size}{The dimension of the conditioning input.}
\item{left_size}{The dimension of the left part of the input split.}
\item{num_bins}{The number of bins in each rational-quadratic spline.}
\item{f_params, g_params}{Conditional networks producing spline parameters.}
\item{tail_bound}{Boundary of the interval transformed by the spline.}
\item{min_bin_width, min_bin_height, min_derivative}{Lower bounds for spline parameters.}
}
\description{
A spline coupling block is a conditional flow inheriting from
\code{nn_conditional_flow()} that applies monotonic rational-quadratic spline
transformations to each side of a split input. It is a more flexible
alternative to \code{nn_affine_coupling_block()}. Values outside
\code{[-tail_bound, tail_bound]} are transformed as the identity.
}
\examples{
library(torch)
flow_model <- nn_spline_coupling_block(2, 0)
x <- torch_randn(10, 2)
y <- flow_model(x)
x_recovered <- flow_model$reverse(y)
}
