% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torchflow-package.R
\docType{package}
\name{torchflow-package}
\alias{torchflow}
\alias{torchflow-package}
\title{torchflow: Conditional Normalizing Flows}
\description{
A package to fit conditional normalizing flows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mbertolacci.github.io/torchflow/}
}

}
\author{
\strong{Maintainer}: Michael Bertolacci \email{m.bertolacci@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
