% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{nn_permutation_flow}
\alias{nn_permutation_flow}
\title{Permutation Flow}
\usage{
nn_permutation_flow(input_size)
}
\arguments{
\item{input_size}{The size of the input to the flow.}
}
\description{
A permutation flow is a conditional flow inheriting from
\code{\link[=nn_conditional_flow]{nn_conditional_flow()}} that permutes the input dimensions. The permutation is
fixed to a random permutation at initialization and does not change. It's
log Jacobian is zero since it is a simple reordering of the input dimensions.
}
\examples{
library(torch)
# Use on its own
permutation_flow <- nn_permutation_flow(10)
input <- torch_randn(10)
output <- permutation_flow(input)
# Use in a more complex conditional flow
flow_model <- nn_sequential_conditional_flow(
  nn_affine_coupling_block(10, 5),
  nn_permutation_flow(10),
  nn_affine_coupling_block(10, 5)
)
}
\seealso{
\code{\link[=nn_conditional_flow]{nn_conditional_flow()}}
}
