% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupling.R
\name{nn_single_coupling_block}
\alias{nn_single_coupling_block}
\title{Single Coupling Block}
\usage{
nn_single_coupling_block(
  input_size,
  conditioning_size = 0,
  left_size = as.integer(input_size\%/\%2),
  transform = "affine",
  params,
  transform_left = TRUE,
  ...
)
}
\arguments{
\item{input_size}{The dimension of the input.}

\item{conditioning_size}{The dimension of the conditioning input.}

\item{left_size}{The dimension of the left part of the input.}

\item{transform}{The transform to apply. Currently \code{"affine"} is supported.
A transform constructor or transform module can also be supplied.}

\item{params}{A conditional network returning the raw transform parameters.}

\item{transform_left}{Whether to transform the left part using parameters
computed from the right part. If \code{FALSE}, the right part is transformed
using parameters computed from the left part.}

\item{...}{Additional arguments passed to the transform constructor.}
}
\description{
A single coupling block splits the input into left and right parts. One part
is left unchanged and is used to compute the parameters for transforming the
other part with a coupling transform.
}
\examples{
library(torch)
coupling <- nn_single_coupling_block(4, transform = "affine")
x <- torch_randn(10, 4)
y <- coupling(x)
x_recovered <- coupling$reverse(y)

}
