% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupling.R
\name{nn_spline_coupling_block}
\alias{nn_spline_coupling_block}
\title{Spline Coupling Block}
\usage{
nn_spline_coupling_block(
  input_size,
  conditioning_size = 0,
  left_size = if (input_size == 1L) 1L else as.integer(input_size\%/\%2),
  ...
)
}
\arguments{
\item{input_size}{The dimension of the input. The input itself is a tensor
with dimensions \verb{[batch_size, input_size]}, or just \verb{[input_size]} if there
is no batch dimension.}

\item{conditioning_size}{The dimension of the conditioning input, which has
the same batch dimensions as the input.}

\item{left_size}{The dimension of the left part of the input.}

\item{...}{Additional arguments passed to the selected coupling block and
spline transform, such as \code{bins}, \code{params} for univariate inputs, or
\code{f_params} and \code{g_params} for multivariate inputs.}
}
\description{
\code{nn_spline_coupling_block()} is a convenience constructor for a
rational-quadratic spline coupling block. It constructs a dual coupling block
when \code{input_size > 1} and a single coupling block when \code{input_size = 1}.
Use \code{\link[=nn_dual_coupling_block]{nn_dual_coupling_block()}} or \code{\link[=nn_single_coupling_block]{nn_single_coupling_block()}} directly to
choose a different transform.
}
\examples{
library(torch)
flow_model <- nn_spline_coupling_block(2, bins = 8)
x <- torch_randn(10, 2)
y <- flow_model(x)
x_recovered <- flow_model$reverse(y)

}
